requirements.txt
Django==1.6.4
Pillow==2.4.0
South==0.8.4
Werkzeug==0.9.4
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
celery==3.1.11
celery-with-redis==3.0
django-celery==3.1.10
django-extensions==1.3.3
django-grappelli==2.5.3
django-mockups==0.4.8
djangorestframework==2.3.13
ipdb==0.8
ipython==2.0.0
kombu==3.0.16
pyreadline==2.0
pytz==2014.2
redis==2.9.1
requirements==0.1
six==1.6.1
sorl-thumbnail==11.12
tablib==0.9.11
wsgiref==0.1.2
